fix: 회원가입 로직 auth로 이동
