+++ src/main/java/com/example/mini_community/service/member/MemberService.java
@@ -1,12 +1,10 @@
 package com.example.mini_community.service.member;
 
 import com.example.mini_community.domain.member.Member;
-import com.example.mini_community.dto.member.CreateMemberRequest;
 import com.example.mini_community.repository.member.MemberRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -15,30 +13,6 @@ public class MemberService {
 
     @Autowired
     private MemberRepository memberRepository;
-
-    private final BCryptPasswordEncoder bCryptPasswordEncoder;
-
-    @Transactional
-    public Member join(CreateMemberRequest req) {
-        Member member = Member.builder()
-                .email(req.getEmail())
-                .password(bCryptPasswordEncoder.encode(req.getPassword()))
-                .nickname(req.getNickname())
-                .profile_img(req.getProfile_img())
-                .build();
-
-        validateDuplicateMember(member);
-
-        return memberRepository.save(member);
-    }
-
-    @Transactional
-    public void validateDuplicateMember(Member member) {
-
-        if (memberRepository.existsByNickname(member.getNickname())) {
-            throw new IllegalStateException("이미 존재하는 닉네임입니다.");
-        }
-    }
 
     @Transactional
     public Member findByEmail(String email) {
